pyproject.toml
[project]
name = "lucidsonicdreams"
version = "0.5.0"
description = "Syncs GAN-generated visuals to music"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Alain Mikael Alafriz", email = "mikaelalafriz@gmail.com" }
]
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["stylegan", "music", "visualization", "gan", "deep-learning", "audio"]

dependencies = [
    "torch>=2.0.0",
    "librosa>=0.10.0",
    "numpy>=1.26.0",
    "moviepy>=1.0.3",
    "Pillow>=10.0.0",
    "tqdm>=4.65.0",
    "scipy>=1.11.0",
    "scikit-image>=0.22.0",
    "pygit2>=1.13.0",
    "gdown>=5.0.0",
    "requests>=2.31.0",
    "soundfile>=0.12.0",
    "imageio>=2.31.0",
    "imageio-ffmpeg>=0.4.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
tensorflow = [
    # For legacy wikiart style support
    "tensorflow>=2.16.0",
]
mega = [
    # For downloading from mega.nz (has older dependency conflicts)
    # Install with: uv pip install mega.py
    "mega.py>=1.0.8",
]

[project.urls]
Homepage = "https://github.com/mikaelalafriz/lucid-sonic-dreams"
Repository = "https://github.com/mikaelalafriz/lucid-sonic-dreams"
Issues = "https://github.com/mikaelalafriz/lucid-sonic-dreams/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lucidsonicdreams"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
